add 2208

+++ LeetCode/2208_将数组和减半的最少操作次数.cpp
@@ -18,6 +18,6 @@ public:
             q.push(top);
             cnt++;
         }
-        return cnt;
+        return static_cast<int>(cnt);
     }
 };
